# documentation/api/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Channel AI Agent API
  description: |
    A scalable multi-channel AI agent service that enables businesses to deploy AI chatbots 
    across WhatsApp, Instagram, and Web Chat platforms.
    
    ## Features
    - Multi-tenant architecture with user_id isolation
    - Support for WhatsApp, Instagram, and Web Chat
    - AI-powered responses using OpenAI
    - Automatic lead extraction and scoring
    - Message delivery tracking
    - Credit-based usage model
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /users/{user_id}/phone_numbers:
    post:
      tags:
        - Users
      summary: Add a phone number
      operationId: addPhoneNumber
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - platform
                - meta_phone_number_id
                - access_token
              properties:
                id:
                  type: string
                platform:
                  type: string
                  enum: [whatsapp, instagram, webchat]
                meta_phone_number_id:
                  type: string
                access_token:
                  type: string
                display_name:
                  type: string
      responses:
        '201':
          description: Phone number added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  timestamp:
                    type: string
                  correlationId:
                    type: string
    
    get:
      tags:
        - Users
      summary: List phone numbers
      operationId: listPhoneNumbers
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
            enum: [whatsapp, instagram, webchat]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Phone numbers retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
  /users/{user_id}/phone_numbers/{phone_number_id}:
    delete:
      tags:
        - Users
      summary: Delete a phone number
      operationId: deletePhoneNumber
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: phone_number_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Phone number deleted successfully
  /users/{user_id}/credits:
    get:
      tags:
        - Users
      summary: Get credit balance
      operationId: getCredits
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credits retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user_id:
                        type: string
                      remaining_credits:
                        type: integer
                      last_updated:
                        type: string
  /users/{user_id}/credits/add:
    post:
      tags:
        - Users
      summary: Add credits
      operationId: addCredits
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Credits added successfully
...